@@ -6,24 +6,23 @@ namespace Dibix.Sdk.CodeGeneration.OpenApi
 {
     internal abstract class ValidationRuleDescriptor<T> : ValidationRuleDescriptor where T : IOpenApiElement
     {
-        public sealed override ValidationRule Create() => new ValidationRule<T>(this.Validate);
+        public sealed override ValidationRule Create() => new ValidationRule<T>(RuleName, Validate);
 
         protected abstract void Validate(IValidationContext context, T target);
     }
 
     internal abstract class ValidationRuleDescriptor
     {
-        private readonly string _ruleName;
+        private string _ruleName;
 
+        protected string RuleName => _ruleName ??= GetRuleName();
         protected abstract string ErrorMessage { get; }
-
-        protected ValidationRuleDescriptor() => this._ruleName = this.GetRuleName();
 
         public abstract ValidationRule Create();
 
         protected virtual string GetRuleName()
         {
-            string name = this.GetType().Name;
+            string name = GetType().Name;
             int suffix = name.LastIndexOf("ValidationRule", StringComparison.Ordinal);
             if (suffix >= 0)
                 name = name.Substring(0, suffix);
@@ -31,6 +30,6 @@ namespace Dibix.Sdk.CodeGeneration.OpenApi
             return name;
         }
 
-        protected void AddError(IValidationContext context) => context.CreateError(this._ruleName, this.ErrorMessage);
+        protected void AddError(IValidationContext context) => context.CreateError(RuleName, ErrorMessage);
     }
 }
