@@ -277,7 +277,7 @@ namespace Dibix.Sdk.Tests.DomainModel
         public int Id { get; set; }
         [DataMember]
         [DefaultValue("DefaultValue")]
-        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
         public string Name { get; set; } = "DefaultValue";
         [DataMember]
         [JsonIgnore]
@@ -288,6 +288,7 @@ namespace Dibix.Sdk.Tests.DomainModel
         public Dibix.Sdk.Tests.DomainModel.Role Role { get; set; } = Dibix.Sdk.Tests.DomainModel.Role.User;
         [DataMember]
         [DateTimeKind(DateTimeKind.Utc)]
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         public System.DateTime? CreationTime { get; set; }
         [DataMember]
         public System.Uri ImageUrl { get; set; }
